--- fe/src/Components/Main/index.js
@@ -1,1 +1,0 @@
-export { Main } from './Main';
